Revert "fixing and refactor the login,  implementing unitary test and adding plugin"
