+++ eslint.config.mjs
@@ -1,7 +1,6 @@
-import pluginJs from '@eslint/js';
-import importPlugin from 'eslint-plugin-import';
-import pluginReact from 'eslint-plugin-react';
-import globals from 'globals';
+import globals from "globals";
+import pluginJs from "@eslint/js";
+import pluginReact from "eslint-plugin-react";
 
 /** @type {import('eslint').Linter.FlatConfig[]} */
 export default [
@@ -9,35 +8,14 @@ export default [
   pluginReact.configs.flat.recommended,
 
   {
-    files: ['**/*.{js,mjs,cjs,jsx}'],
+    files: ["**/*.{js,mjs,cjs,jsx}"],
     languageOptions: { globals: globals.browser },
-    plugins: { react: pluginReact, import: importPlugin },
+    plugins: { react: pluginReact },
     rules: {
       ...pluginReact.configs.recommended.rules,
-      'react/prop-types': 'off',
-      'react/react-in-jsx-scope': 'off',
-      'import/order': [
-        'error',
-        {
-          groups: [
-            'builtin',
-            'external',
-            'internal',
-            ['sibling', 'parent', 'index'],
-          ],
-          'newlines-between': 'always',
-          alphabetize: {
-            order: 'asc',
-            caseInsensitive: true,
-          },
-        },
-        {
-          env: {
-            jest: true,
-          },
-        },
-      ],
+      "react/prop-types": "off",
+      "react/react-in-jsx-scope": "off"
     },
-    settings: { react: { version: 'detect' } },
+    settings: { react: { version: "detect" } },
   },
 ];
