+++ package.json
@@ -11,8 +11,8 @@
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.9.0",
     "json-server": "^1.0.0-beta.3",
-    "react": "^19.2.0",
-    "react-dom": "^19.2.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "react-hook-form": "^7.56.4",
     "react-icons": "^5.5.0",
     "react-router-dom": "^7.6.2",
@@ -51,11 +51,9 @@
   "devDependencies": {
     "@eslint/js": "^9.26.0",
     "eslint": "^8.57.1",
-    "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-react": "^7.37.5",
     "eslint-plugin-react-hooks": "^5.2.0",
     "globals": "^16.1.0",
-    "prettier": "^3.5.3",
-    "react-test-renderer": "^19.2.0"
+    "prettier": "^3.5.3"
   }
 }
