+++ src/App.js
@@ -1,21 +1,22 @@
-import { useEffect, useState } from 'react';
+import { useState, useEffect } from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Swal from 'sweetalert2';
-
-import { AuthContextProvider } from './context/auth.jsx';
 import useWakeUpAPI from './hooks/useWakeUpAPI.js';
-import Modal from './components/Modal';
-import UserForm from './components/Forms/UserForm';
-import Header from './components/Header';
+
 import { Home } from './pages/home';
-import ListProducts from './pages/listProducts';
 import ProductsRegister from './pages/register/products';
+import ListProducts from './pages/listProducts';
+
+import Swal from 'sweetalert2';
 
 import './styles/reset.sass';
 
 function App() {
-  const [showModal, setShowModal] = useState(false);
-  const [isFormRegister, setIsFormRegister] = useState(false);
+  const [currentUser, setCurrentUser] = useState({
+    status: false,
+    email: '',
+    name: '',
+  });
+
   const statusAPI = useWakeUpAPI();
 
   useEffect(() => {
@@ -52,31 +53,25 @@ function App() {
   }, []);
 
   return (
-    <>
-      <Router>
-        <AuthContextProvider>
-          <Header
-            setShowModal={setShowModal}
-            setIsFormRegister={setIsFormRegister}
-          />
-
-          <Routes>
-            <Route path="/" element={<Home statusAPI={statusAPI} />} />
-            <Route path="/register/products" element={<ProductsRegister />} />
-            <Route path="/list/:category" element={<ListProducts />} />
-          </Routes>
-
-          {showModal && (
-            <Modal setShowModal={setShowModal}>
-              <UserForm
-                setShowModal={setShowModal}
-                isFormRegister={isFormRegister}
-              />
-            </Modal>
-          )}
-        </AuthContextProvider>
-      </Router>
-    </>
+    <Router>
+      <Routes>
+        <Route
+          path="/"
+          element={
+            <Home
+              currentUser={currentUser}
+              setCurrentUser={setCurrentUser}
+              statusAPI={statusAPI}
+            />
+          }
+        />
+        <Route path="/register/products" element={<ProductsRegister />} />
+        <Route
+          path="/list/:category"
+          element={<ListProducts currentUser={currentUser} />}
+        />
+      </Routes>
+    </Router>
   );
 }
