+++ src/components/Aside/index.jsx
@@ -1,8 +1,8 @@
-import { useState } from 'react';
-import { FaAngleLeft, FaAngleRight } from 'react-icons/fa';
-
-import useWindowWidth from '../../hooks/useWindowWidth';
 import Button from '../Button';
+import { useState } from 'react';
+import useWindowWidth from '../../hooks/useWindowWidth';
+
+import { FaAngleLeft, FaAngleRight } from 'react-icons/fa';
 
 import './styles.sass';
