+++ src/components/Footer/index.jsx
@@ -1,10 +1,9 @@
+import "./styles.sass";
 import {
   FaPhone,
   FaRegEnvelope,
   FaLocationDot
 } from "react-icons/fa6";
-
-import "./styles.sass";
 
 const Footer = () => {
   return (
