+++ src/components/Header/index.jsx
@@ -31,7 +31,6 @@ const Header = ({ currentUser, setShowModal, setIsFormRegister }) => {
   const [isOpen, setIsOpen] = useState(false);
   const navigate = useNavigate();
   const width = useWindowWidth();
-  const { currentUser } = useContext(AuthContext);
 
   const handleCategoryClick = useCallback(
     (category) => {
