+++ src/components/Modal/index.jsx
@@ -1,6 +1,6 @@
-import { RiCloseLargeLine } from "react-icons/ri";
-
 import Button from '../Button';
+
+import { RiCloseLargeLine } from "react-icons/ri";
 
 import './styles.sass';
