--- src/context/auth.jsx
@@ -1,41 +1,0 @@
-import { createContext, useState } from 'react';
-
-import { fetchLogin } from '../services/loginService.js';
-
-export const AuthContext = createContext();
-
-export const AuthContextProvider = ({ children }) => {
-  const [currentUser, setCurrentUser] = useState({
-    status: false,
-    email: '',
-    name: '',
-  });
-
-  const handleLogin = async (email, password) => {
-    const users = await fetchLogin();
-
-    if (!users) {
-        return 'errorServer';
-    }
-
-    const user = users.find(
-      (user) => user.email === email && user.password === password
-    );
-
-    if (user) {
-      setCurrentUser({
-        status: true,
-        email: email,
-        name: user.username,
-      });
-
-      return 'ok'
-    } else {
-      return 'falied'
-    }
-  };
-
-  return (
-    <AuthContext.Provider value={{ currentUser, setCurrentUser, handleLogin }}>{children}</AuthContext.Provider>
-  );
-};
