+++ src/hooks/useWakeUpAPI.js
@@ -1,5 +1,4 @@
 import { useState, useEffect } from 'react';
-
 import { api } from '../services/api.js';
 
 const useWakeUpAPI = () => {
