+++ src/index.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-
 import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
