+++ src/pages/listProducts/index.jsx
@@ -1,8 +1,10 @@
-import { useEffect, useState, useCallback } from 'react';
-import { useParams } from 'react-router-dom';
-
+import Header from '../../components/Header';
 import Aside from '../../components/Aside';
 import CardProduct from '../../components/CardProduct';
+
+import { useParams } from 'react-router-dom';
+import { useEffect, useState, useCallback } from 'react';
+
 import { api } from '../../services/api.js';
 import {
   filterProductsByCategory,
@@ -12,7 +14,7 @@ import { removeHyphen } from '../../utils/textFormatter.js';
 
 import './styles.sass';
 
-const ListProducts = () => {
+const ListProducts = ({ currentUser }) => {
   const { category } = useParams();
   const [products, setProducts] = useState([]);
   const [selectedTag, setSelectedTag] = useState('');
@@ -50,6 +52,7 @@ const ListProducts = () => {
 
   return (
     <>
+      <Header currentUser={currentUser}/>
       <div className="container">
         <Aside
           title={category.includes('-') ? removeHyphen(category) : category}
