+++ src/pages/register/products/index.jsx
@@ -1,11 +1,13 @@
-import { yupResolver } from '@hookform/resolvers/yup';
-import { useCallback } from 'react';
-import { useForm } from 'react-hook-form';
-import Swal from 'sweetalert2';
-import * as yup from 'yup';
-
 import Button from '../../../components/Button';
 import Input from '../../../components/Input';
+
+import Swal from 'sweetalert2';
+
+import { useCallback } from 'react';
+import { useForm } from 'react-hook-form';
+import { yupResolver } from '@hookform/resolvers/yup';
+import * as yup from 'yup';
+
 import { createProduct } from '../../../services/productService.js';
 
 import './styles.sass';
